Print seqno and ackno for debugging
Also simplifies the filter as the super complex one from man pcap-filter
didn't seem to actually filter anything. May as well take all the tcp
packets.

Plan now is to work through the example dump in wireshark with a pen and
paper to work out a sesnible algorithm for reconstructing the http
stream. It would help if I understood networking better :sweat_drops: